Factor out logger setup
